add tests

+++ src/pages/api/session.js
@@ -1,11 +1,12 @@
-import { ErrorHandler } from '../../utils/ErrorHandler'
+import {
+	ErrorHandler,
+	HttpUnauthorizedError,
+} from '../../utils/ErrorHandler'
 import { ApiHelper } from '../../utils/ApiHelper'
 import {
 	HTTP_METHOD,
 	HTTP_STATUS,
 } from '../../utils/constants'
-import { CryptoHelper } from '../../utils/CryptoHelper'
-import { UserDAO } from '../../dao/UserDAO'
 import { CookieHelper } from '../../utils/CookieHelper'
 
 export default async function handler(req, res) {
@@ -14,46 +15,27 @@ export default async function handler(req, res) {
 		ApiHelper.checkHttpMethod(req, allowed_methods)
 
 		if (req.method === HTTP_METHOD.GET) {
-			// check credentials
-			const { session_token } = req.cookies
-			if (!session_token) {
-				res.status(HTTP_STATUS.NO_CONTENT).end()
-				return
+			try {
+				const user = await ApiHelper.checkSessionToken(req)
+				res.status(HTTP_STATUS.OK).json({
+					username: user.username,
+				})
+			} catch (err) {
+				if (err instanceof HttpUnauthorizedError) {
+					// the session cannot be used,
+					// so erase it from browser
+					const cookie_obj = {
+						session_token: '',
+					}
+					CookieHelper.deleteCookie(res, cookie_obj)
+
+					res.status(HTTP_STATUS.NO_CONTENT).end()
+				} else {
+					throw err
+				}
 			}
-
-			const decrypted = JSON.parse(CryptoHelper.decrypt(session_token))
-			const { username } = decrypted
-			if (!username) {
-				res.status(HTTP_STATUS.NO_CONTENT).end()
-				return
-			}
-
-			const filter = { username: username }
-			const user = await UserDAO.get(filter, ['username'])
-			if(!user) {
-				res.status(HTTP_STATUS.NO_CONTENT).end()
-				return
-			}
-
-			// update user record
-			await UserDAO.upsert(filter, {
-				last_active_at: new Date(),
-			})
-
-			// done
-			res.status(HTTP_STATUS.OK).send({
-				username: username,
-			})
 		}
 	} catch (err) {
-		// the session cannot be used,
-		// so erase it from browser
-		const cookie_obj = {
-			session_token: '',
-		}
-		CookieHelper.deleteCookie(res, cookie_obj)
-
-		// handle error
 		ErrorHandler.handleApiError(err, req, res)
 	}
 }
